Update toolchain to Rust 2024 edition (#269)
* In progress error update.

* In progress error update.

* In progress toolchain update.

* In progress error update.

* In progress error update.

* In progress error update.

* Toolchain update working.

* Toolchain update working.

* Toolchain update working.

* In progress toolchain update.

* Copy libclang-rt during toolchain install.

* Copy libclang-rt during toolchain install.

* Copy llvm toolchain during bootstrap.

* Copy llvm toolchain during bootstrap.

* mlibc compiling.

* Fix libc link.

* Add libc test.

* Fix mlibc build dir.

* Fix mlibc build dir.

* Fix mlibc build dir.

* Fix mlibc build dir.

* Fix aarch64 build.

* Fix build.

* Fix llvm build.

* Fix llvm build.

* Fix build.

* Fix build.

* Fix build.

* Fix(ing) CI scripts.

* Fix(ing) CI scripts.

* Fix(ing) CI scripts.

* Fix(ing) CI scripts.

* Fix(ing) CI scripts.

* Fix(ing) CI scripts.

* Fix(ing) CI scripts.

* Get hello world in C working.

* Add lwext4.

* Fix(ing) CI scripts.

* Add security error codes.

* Update docs.

* Fix(ing) CI scripts.

* Add security error codes.

+++ src/lib/twizzler/src/ptr/invariant.rs
@@ -65,9 +65,9 @@ impl<T: Invariant> InvPtr<T> {
         Self::from_raw_parts(0, 0)
     }
 
-    pub fn from_raw_parts(idx: u64, offset: u64) -> Self {
+    pub fn from_raw_parts(idx: u32, offset: u64) -> Self {
         Self {
-            value: (idx << 48) | offset,
+            value: ((idx as u64) << 48) | offset,
             _pse: PhantomStoreEffect,
             _pd: PhantomData,
         }
